chore(eq): Implement an optimized equality theory that partition variables into potentially unifiable ones.
